Extract SlideShape types and props into separate pure-TS file
Match the existing ThemeImageShape.ts pattern by moving SlideShapeType,
SlideShape type, and shape props into SlideShape.ts, keeping
SlideShapeUtil.tsx focused on the React rendering logic.
